refactor: widen PDF integer types to eliminate narrowing casts
XrefEntry.gen, ObjRef.obj/gen, HashMap key, parseUint return, xref
stream /W widths and /Index pairs all widened to u64/usize. Removes
7 narrowing casts entirely. 15 remaining std.math.cast are legitimate
(i64→usize for negative checks, u64→usize for 32-bit compat).

+++ src/core/parser_pdf.zig
@@ -342,7 +342,7 @@ pub fn parse(allocator: Allocator, content: []const u8, path: []const u8) !Docum
 }
 
 /// Recursively collect text spans from a page tree node (Pages or Page).
-fn collectPageSpans(allocator: Allocator, ctx: *PdfContext, obj_num: u32, spans: *std.ArrayList(TextSpan), page_counter: u32) PdfError!void {
+fn collectPageSpans(allocator: Allocator, ctx: *PdfContext, obj_num: u64, spans: *std.ArrayList(TextSpan), page_counter: u32) PdfError!void {
 	const obj = (ctx.getObject(obj_num) catch return) orelse return;
 	defer pdf_objects.freePdfValue(allocator, obj);
